requirements.txt
python-dotenv
pathlib3x
google-cloud-vision
